requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
anyio==4.13.0
attrs==26.1.0
backtrader==1.9.78.123
catboost==1.2.10
ccxt==4.5.49
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
coincurve==21.0.0
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
fonttools==4.62.1
frozenlist==1.8.0
graphviz==0.21
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imbalanced-learn==0.14.1
joblib==1.5.3
kiwisolver==1.5.0
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.11
MarkupSafe==3.0.3
matplotlib==3.10.8
multidict==6.7.1
narwhals==2.19.0
numba==0.61.2
numpy==2.2.6
nvidia-nccl-cu12==2.29.7
optuna==4.8.0
packaging==26.1
pandas==3.0.2
pandas-ta==0.4.71b0
pillow==12.2.0
plotly==6.7.0
propcache==0.4.1
psutil==7.2.2
pycares==5.0.1
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.7
PyYAML==6.0.3
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
sklearn-compat==0.1.5
SQLAlchemy==2.0.49
ta==0.11.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
xgboost==3.2.0
yarl==1.23.0
